requirements.txt

langchain
langchain-openai
google-generativeai 
langchain-google-genai 
langchain-core
pypdf
Pillow
pytesseract
pandas
streamlit
python-dotenv

